# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: EMS Logistics API
  version: 1.0.0
  description: Contract-first OpenAPI specification for EMS Logistics System.
  license:
    name: Proprietary
    url: https://ems-logistics.local/license
servers:
  - url: /api
    description: Local and deployed API base path
...
paths:
...
  /trip-legs/{tripLegId}:
    parameters:
      - name: tripLegId
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Trips]
      operationId: getTripLegById
      summary: Get trip leg
      responses:
        "200": { description: Trip leg, content: { application/json: { schema: { $ref: "#/components/schemas/TripLegResponse" } } } }
  /trip-legs/{tripLegId}/assign-booking:
    parameters:
      - name: tripLegId
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Trips]
      operationId: assignBookingToTripLeg
      summary: Assign booking to trip leg
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AssignBookingToTripLegRequest" } } } }
      responses:
        "200": { description: Booking assigned, content: { application/json: { schema: { $ref: "#/components/schemas/TripLegResponse" } } } }
  /trip-legs/{tripLegId}/dispatch:
    parameters:
      - name: tripLegId
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Trips]
      operationId: dispatchTripLeg
      summary: Dispatch trip leg
      responses:
        "200": { description: Trip leg dispatched, content: { application/json: { schema: { $ref: "#/components/schemas/TripLegResponse" } } } }
  /trip-legs/{tripLegId}/complete:
    parameters:
      - name: tripLegId
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Trips]
      operationId: completeTripLeg
      summary: Complete trip leg and update vehicle station
      responses:
        "200": { description: Trip leg completed, content: { application/json: { schema: { $ref: "#/components/schemas/TripLegResponse" } } } }
  /trip-legs/{tripLegId}/emergency-override:
    parameters:
      - name: tripLegId
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Trips]
      operationId: emergencyOverrideTripLeg
      summary: Apply emergency vehicle/driver override
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/EmergencyTripOverrideRequest" } } } }
      responses:
        "200": { description: Override applied, content: { application/json: { schema: { $ref: "#/components/schemas/TripLegResponse" } } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    TripLegStatus:
      type: string
      enum: [planned, dispatched, arrived, completed, cancelled]
...
    TripLeg:
      type: object
      required: [id, masterTripId, sequenceNo, route, originStation, destinationStation, plannedDepartureAt, actualDepartureAt, actualArrivalAt, status, overrideVehicle, overrideDriver, overrideReason, overrideApprovedByUserId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        masterTripId: { type: integer }
        sequenceNo: { type: integer }
        route:
          type: object
          required: [id, name]
          properties: { id: { type: integer }, name: { type: string } }
        originStation:
          type: object
          required: [id, name]
          properties: { id: { type: integer }, name: { type: string } }
        destinationStation:
          type: object
          required: [id, name]
          properties: { id: { type: integer }, name: { type: string } }
        plannedDepartureAt: { type: string, format: date-time, nullable: true }
        actualDepartureAt: { type: string, format: date-time, nullable: true }
        actualArrivalAt: { type: string, format: date-time, nullable: true }
        status: { $ref: "#/components/schemas/TripLegStatus" }
        overrideVehicle:
          nullable: true
          type: object
          required: [id, vehicleNumber]
          properties: { id: { type: integer }, vehicleNumber: { type: string } }
        overrideDriver:
          nullable: true
          type: object
          required: [id, name]
          properties: { id: { type: integer }, name: { type: string } }
        overrideReason: { type: string, nullable: true }
        overrideApprovedByUserId: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssignBookingToTripLegRequest:
      type: object
      required: [bookingId]
      properties:
        bookingId: { type: integer }
        allocatedWeightTons: { type: number, nullable: true }
    EmergencyTripOverrideRequest:
      type: object
      required: [reason]
      properties:
        vehicleId: { type: integer, nullable: true }
        driverId: { type: integer, nullable: true }
        reason: { type: string }
        approvedByUserId: { type: integer, nullable: true }
...
    TripLegResponse:
      type: object
      required: [data, message]
      properties: { data: { $ref: "#/components/schemas/TripLeg" }, message: { type: string } }
...
security:
  - bearerAuth: []